+++ CHANGELOG.md
@@ -1,2 +1,6 @@
 # Change Log
-All notable changes to the "php-resolver" extension will be documented in this file.
+
+## 1.0.2
+- Improved searching
+- Rewrite importing logic and class name checking
+- Added file-size in status bar as part of the extension